Remove CSRF token from search form
The CSRF tag adds a random token[1] that ensures that forms cannot be
submitted directly without the user visiting the app first. This is
important when taking destructive actions or submitting data.

In the case of search we want the opposite though: users should be able to
bookmark searches and share working links. CSRF tokens should therefore
not be added to the query parameters.

[1]. https://docs.djangoproject.com/en/5.0/ref/csrf/