Add javascript for domain filter widget
Domain will have top level and subdomain selections, and work similarly
to "Topic/Sub-Topic" on GOV.UK search.

The form will submit domain and subdomain as separate fields, so we
need to combine them on the backend.

If javascript is not enabled, then the subdomain field is not displayed
and it will work as before.

This is tested using jest and jest-dom.