Introduce table details page (#136)
* Extract package tests into modules

* Introduce dataset template

Start with description and schema. Will add the rest later

* Replace back link with breadcrumb

This presents every search result as belonging to "Search"

This link remembers the last valid search made via the session, so the
user doesn't lose context when navigating back and forth. However, if
starting a new session via a deep link, the user can still go back
to the search page.

This commit also fixes the layout for this template as the full width
container and main element were repeated in the base template and the
details template, and we were missing a column div, which messed up
the margins of the h1 header.

* Populate dataset details page