Demote headings appearing in markdown (#166)
Currently, datahub can store arbitrary markdown in description fields.
When we render this, it interferes with the normal heading structure of
the page, because we have H1s following H3s etc.

This markdown extension fixes this problem by automatically demoting
headings a number of levels depending on the context we are rendering
in.

This solution isn't perfect, because sometimes layout headings use a
smaller font size than the default for the corresponding level. Markdown
rendered headings always use the default defined by the design system,
so in such cases we may need to skip a level to get consistent sizing.