Use a regular list for metadata, not a summary list
The GOV.UK summary list component introduces quite a lot
of additional whitespace between items, since it's designed
for careful review of user-submitted data.

In our design, the metadata is much less prominent, and is
there for users to skim over quickly. We therefore want
to condense this component into a smaller space.