Stop showing "TBC" as "Database Name" in search results (#236)
Drop database name

This has been there since the beginning of the prototype, but it's not
included in the most recent designs, and it's hardcoded to TBC.

Since this is currently not adding any value, we may as well get rid of
it. In the next iteration we may change how we display search result
titles and add back in the fully qualified name or database name,
but we can do this once we know what we want to show.