Workaround for skip link issues (#246)
The $govuk-global-styles setting applies the govuk-focus-text mixin to
all links, which removes underline and adds box-shadow.  But the skip
link component removes box shadow from the focus state, without adding
back in the underline.  This has the effect that skip links are not
underlined when $govuk-global-styles is true.

This workaround reenables the underline. This commit may be reverted
when we pull a new version of the design system with a fix in.