Add nltk stemming to make the highlighting more reliable
Currently we only highlight words when a search term is a prefix of
the word. If we first reduce the search term to its stem, we are able
to match variants of the word, e.g. I can search for the plural and
still see matches for the singular.

This uses the PorterStemmer algorithm from nltk. This algorithm may
reduce words with a different meaning but common derivation, but it's
simple enough that don't need to anything more complicated than the
regex matching we were doing before.

For more info on stemming vs lemmatization, see
https://www.datacamp.com/tutorial/stemming-lemmatization-python