Better highlighting in descriptions (#259)
* Improve match highlighting for multiple word queries

Highlight at a word level if some words in the query match the
description.

* Fix type error

* Add a template filter for smarter search snippets

This filter will truncate to a fixed number of characters, but
will try to include <mark></mark> sections in the output.

* Increase snippet length to avoid losing the start of paragraphs

This gives a better snippet in the "Releases in error" example:

> A prisoner is released in error if they are wrongly discharged from a prison or court when they should have remained in custody, where the prisoner has not deliberately played a part in the error (i.e. the prisoner had no intent of escaping).
>
> Examples include misplaced warrants for imprisonment or r…

The length calculation is a bit off here because this metadata contains
extra HTML we have picked up from the justice data API. The markdown
process handles this sensibly but it means we overestimate the text
length compared to plain text or markdown.

* Add nltk stemming to make the highlighting more reliable

Currently we only highlight words when a search term is a prefix of
the word. If we first reduce the search term to its stem, we are able
to match variants of the word, e.g. I can search for the plural and
still see matches for the singular.

This uses the PorterStemmer algorithm from nltk. This algorithm may
reduce words with a different meaning but common derivation, but it's
simple enough that don't need to anything more complicated than the
regex matching we were doing before.

For more info on stemming vs lemmatization, see
https://www.datacamp.com/tutorial/stemming-lemmatization-python

* Fixup