PoC for a more/less toggle component
This demonstrates a component for hiding text at a defined cutoff point,
and allowing it to be revealed by a button.

This works with plain text or html containing inline elements, but it
doesn't work reliably with markdown formatted text.

If we render markdown before partitioning the text, then truncate_chars_html will
close block level elements, and the ellipsis won't be at the end of the
string.

This breaks the computation of prefix and remainder, and introduces
artificial line breaks between the two.

On the other hand if we render the markdown after partitioning the text,
then this tag needs to be markdown-aware to avoid breaking text in
places that would interfere with the markdown rendering - e.g. partway
through a markdown link. So the simple approach of `truncatechars_html`
is not good enough.