Mitigation for overflowing table (desktop)
Since the column descriptions may contain arbitrary markdown text,
we have examples where this table stretches beyond the grid column.

It's possible we will be able to eliminate this by improving the
content, but to mitigate against future issues, we can give cells
in this column a max width so they can't grow indefinitely.