Stable tables (#267)
* Mitigation for overflowing table (desktop)

Since the column descriptions may contain arbitrary markdown text,
we have examples where this table stretches beyond the grid column.

It's possible we will be able to eliminate this by improving the
content, but to mitigate against future issues, we can give cells
in this column a max width so they can't grow indefinitely.

* Include markdown extensions

Render fenced code blocks when rendering markdown.

This is not very pretty at the moment, but it's better than not
rendering it, and datahub supports this syntax so we may as well
support it as well. Code snippets may be useful in some cases, e.g.
example SQL queries.

I've decided not to to include the `tables` extension here, since
datahub doesn't support that, and we should be able to scrub these
from the metadata we're currently ingesting.