Add a details component with search query tips
Provide some query examples highlighting some of the supported syntax,
and link to the full documentation in Datahub.

This should help mitigate the sometimes unintuitive default behaviour of
requiring that ALL search terms match.

This behaviour is how Datahub have configured search, and to change it
we would have to customise the Elasticsearch query.