Consistently parse names from datahub (#4244)
* Consistently parse names from datahub

Previously, we were parsing names inconsistently, and there were edge
cases where we would get unexpected None values or KeyErrors.

Datahub has 4 possible names that might be set

- name: Unique guid for dataset No longer to be used as the Dataset display name. Use properties.name instead
- properties.name
- properties.displayName
- properties.qualifiedName

These are used inconsistently by different ingestion sources. If we have
displayName / qualifiedName it's unambiguous, but the top level name can
be either.

We expose 3 values:
- `displayName` and `qualifiedName` align with the corresponding datahub
names, and fall back to the generic name if not set.
- `name` is probably redundant now, but I've kept it in to avoid breaking
things further.

* Lint

* Add missing __init__.py

* Disable pylint warnings for missing docstrings

Not sure why this has suddenly started being checked, but we don't care
right now. We will soon move this to a different repo and settle on
a single linter.