Coerce descriptions and custom properties to string/empty string (#4256)
* Coerce column description to string

In some cases we column description is set to None, but we have defined
it as a non-nullable string. For simplicity, we can treat None as
empty string, as this is what we'll actually display in the frontend.

Resolves https://github.com/ministryofjustice/find-moj-data/issues/296

* Simplify type of list field

* Simplify types

Custom properties values are always String in datahub, so using string
throughout will make our lives easier. We can use empty string to
indicate missing values here.
See https://datahubproject.io/docs/graphql/objects/#custompropertiesentry

dpia_required is a special case - it's still a string in datahub but we
want it to be a boolean when we use it - since "False" is truthy
in python.

When setting custom properties, we take the string representation. This
will serialize booleans to 'True' and 'False' which will fail pydantic's
validation, so we need to explicitly convert this back to a boolean.
https://docs.pydantic.dev/2.0/usage/types/booleans/

If dpia_required is unset, we should just pass empty string, rather than
"None".