Fix rendering bug with Matched fields
If matched fields containes a thing we don't recognise (e.g. a custom
property pulled from DBT) then we were treating it as an empty string,
resulting in a matched fields value that has a comma at the beginning.

To avoid this, change the lookup to ignore any keys that are not in
the dictionary.