Make search a search input, and style its close button
This allows searches to be cancelled by clicking the cross
or pressing escape.

This should be used both for the plain styled search on the glossary
page, and the combined input + icon variant on the search page.

This is loosely based on the GOV.UK component: https://components.publishing.service.gov.uk/component-guide/search
However, I've used the close icon already present in the MOJ design
system.