Split up selenium test cases, and add homepage tests
The selenium test was becoming long and unmanagable, and it wasn't easy
to get a high level picture of what we were testing.

This commit splits them into a few different groups
- testing interacting with search results to view the details page
- testing the behaviour of search controls
- testing navigating to pages using the top level nav

and adds tests for searching and browsing directly from the homepage.

This creates a bit of duplication, but if it becomes a problem we can
factor out a base class.