Add a little check to make sure the pre-commit hooks ran as expected
I've made this into a comment rather than a failing check, because
I've noticed instances where the pre-commit hook doesn't do the right
thing.

For example, djhtml refuses to process templates/navigation.html,
because it contains a long line. This is because the crown logo SVG from
the header contains a very long path attribute.