Small fixes (#571)
* Replace YES and NO with Yes and No

* Handle cases where properties and/or name is missing

* Make chart links open in new tab by default

Previously this opened in the current tab, but we received feedback that
this was not what the user expected. Also, this caused confusion in a
prior research round, because both Justice Data and Find MOJ Data look
the same, but they are not currently designed to work as one site.
Opening in a new tab makes it clearer that these are separate services.

I've included (open in new tab) text but made it visually part of the
header rather than part of the link itself, since we have chosen to
render the full URL as the link text, which looks a bit weird with stuff
in parentheses. When using a screenreader, "(open in new tab)" is still
part of the link text.