docs: update readme
Add some notes about configuration for other orgs who may wish to reuse
this frontend. This is just a starting point for now.

As we speak more with the Scottish Government team we should get a
better understanding of the pain points and we can figure out a release
process that minimises breakages.