feat: streamline deploy pipeline
The current pipeline has a couple of problems:

1. We use github's environment protection, but we access environment
   scoped variables in multiple jobs per environment, so we are prompted
   twice for approval to continue.

2. We build the docker image once per environment, so we can't
   guarantee the integrity of the one we end up with in prod.

To address this I have moved the ECR variables to repository level, so
they can be used without setting an `environment`.

In deploy-staged.yml, we now build the image once, without requiring an
extra approval step. Since we have repositories for each environment,
the image is pushed to the next environment as soon as we have
successfully deployed to the preceding environments.

I've also removed `deploy-to-dev.yml`, which triggered for every PR, because

1. It required us to have an otherwise unnecessary approval gate on
   dev, to prevent us unintentionally deploying on top of a branch
   somebody else is testing
2. It made it impossible to distinguish between "tests are still
   running" and "tests passed but the branch wasn't deployed" when
   looking at a list of PRs
3. We have a manually triggered `deploy-dev-from-branch.yml` which
   deploys a branch to dev without these problems.