feat: Adds govuk '< Back' button to the cookies page
The cookies page is viewable from any page in the application, but is likely not an endpoint for users, so they will want to be able to navigate back to their previous page.

The cookie view has been added to `home/views.py`, and the referer URL is validated to ensure the URL is one of the trusted domains (i.e. is within `settings.CSRF_TRUSTED_ORIGINS`).

localhost has been added to `settings.CSRF_TRUSTED_ORIGINS` when running with `settings.DEBUG == True`