Disable datahub telemetry and force debug off (#857)
Datahub has some telemetry. This seems to be disabled on the server
side, but unless we set this flag, the GraphQL client will issue a
request to the server asking for a clientId. This always returns a
404 on our deployment.

This request adds some delay to response times in Find MoJ data, and this
seems is made worse by us instantiating the client in multiple places.

I've also hardcoded DEBUG to 0 since we will never have this on outside of
local development.