Hide schema section for performance hub metrics and display caveats (#943)
Add custom template for Performance Hub Metrics

Since performance hub metadata is not in prod yet and are subject to change,
I've not implemented Metrics as a distinct entity type that can be filtered
on. Instead,

1. expose a subtype attribute that determines what template to render
2. conditionally add extra caveats to metrics coming from
   performance-hub, to reflect their current state

This will need to be reviewed again if we decide to move forward with
performance hub metadata.