feat: don't ask users for their email
Instead of requiring the reporter to enter their email, we can just use
the email of the currently signed in user.

Instead, just ask if they would like to be emailed.

I'm using a TypedChoiceField instead of a BooleanField, because
BooleanField has weird `required` behaviour. The docs encourage you
to set `required=False` if you need to be able to submit a False value,
but this means you can't validate that a value was set if you use a
RadioWidget.