fix: refer to data owners at custodians
This is a temporary workaround. Once the next version of Datahub is
released, we will be able to reenable the `meta_mapping` in the dbt
ingestion, which will allow us to store Data custodians as Data
custodians.

At this point we will be able to distinguish between
different kinds of owners in Datahub, and we can remove this workaround.

The ability to distinguish between owners and custodians is not
essential at the moment, but it's desirable, because the data improvement
team are interesting data owner/data steward/data custodian roles, and
it's very likely that data owners and stewards will also interact with
the Find MoJ data and/or datahub.