Enable trivy image scan on PRs
For now, do not fail the build if there are any vulnerabilities
reported. The vulnerabilities will appear in the code scanning
page in the repo security tab, and github will warn when introducing
new vulnerabilities.

Later, we could choose to fail on any critical vulnerabilities,
as in this workflow:
https://raw.githubusercontent.com/ministryofjustice/analytical-platform-ingestion-notify/refs/heads/main/.github/workflows/scan-image.yml

.trivyignore.yaml can optionally be used if we want to ignore
vulnerabilities until a certain date.