Docker file refactor (#1066)
* refactor(Dockerfile): copy less files to node builder

This reduces the build time from 89s to 70s. Final runtime image is
unchanged.

* refactor(Dockerfile) move compilemessages to runtime image

compilemessages depends on the whole Django app, whereas the poetry
install doesn't, so moving this into the runtime image should simplify that
initial build stage and makes it more cachable.

This doesn't affect the overall build time without cache. The gettext
package adds about 8mb to the final image.

* refactor(Dockerfile) limit runtime image

This refactor simplifies the runtime image stage and limits
the files copied into the image.

This halves the image size from 505MB to 221MB, and reduces the build
time from 70s to 56s when cache is not used.

* minor comment changes