refactor: rename timestamps (#1110)
* refactor: rename timestamp parse functions

There are various datetimes returned by Datahub, and it's easy to
confuse them. It seems like we are not always using the right one, so as
a first step I'm renaming all the parse functions to clearly describe
what they are actually returning.

* refactor: rename timestamp fields on Entity class

We have several timestamps that are easily confused. To try and make the
code clearer, I'm defining them for all entities with explicit names
that correspond to how they are stored in Datahub.