fix: stop showing subject area as a matched field
Datahub returns us a matchedFields object, containing a list of fields
which matched the query.

We used to look for the presence of domain here and display "Subject area"
if that is the case, however going forwards we will decouple Datahub
domains from subject areas so this is not correct.

Removing this means we will not show subject area in the list of matched
fields, even if the result is tagged to the subject area. This is
intentional -- preserving this behaviour would require some extra logic
to work out whether the tag that matched was a subject area tag, and I
don't think it's its needed, since the user can clearly see when a
result has matched on subject area (we always display the subject area).