Subject area tag cleanup work (#1246)
* fix: use cache versioning

Since the cache stores arbitrary objects, loading from cache
can fail if classes change and the old values are not purged
from the cache.

Setting a version sidesteps that issue as we can change the
version when making a breaking change like this
https://docs.djangoproject.com/en/5.1/topics/cache/#cache-versioning

* refactor: rename template variable

* refactor: use tag URN in search URL query paramaters

* refactor: remove references to domains

* refactor: remove remaining subdomains code

we took out subdomains already, and we're not intending to add them back
in the immediate future.

* refactor: change form field from "domain" to "subject_area"

* fixup selenium tests

* fix: stop showing subject area as a matched field

Datahub returns us a matchedFields object, containing a list of fields
which matched the query.

We used to look for the presence of domain here and display "Subject area"
if that is the case, however going forwards we will decouple Datahub
domains from subject areas so this is not correct.

Removing this means we will not show subject area in the list of matched
fields, even if the result is tagged to the subject area. This is
intentional -- preserving this behaviour would require some extra logic
to work out whether the tag that matched was a subject area tag, and I
don't think it's its needed, since the user can clearly see when a
result has matched on subject area (we always display the subject area).

* feat: handle multiple subject areas assigned to the same entity

Previously an entity could have only a single subject area. Now we've
switched to using tags, there can be more than one. In this case we
should display them as a comma separated list.

* refactor: remove domain references from tests

* refactor: remove DomainRef class

* refactor: remove unused create_domain method

* refactor: remove unused domain class

* refactor: remove unused domain fields from GraphQL queries

* chore: remove duplicate ESDA script

The other one is in the data-catalogue-metadata repo. I will merge any
changes into that one so we don't lose anything.

* refactor: don't refer to TOP_LEVEL, there's only one level now

* refactor: remove redundant and confusing class

There are now two classes called SubjectAreaTaxonomy.  One lives in the
datahub_client package, and the other is a holdover from when we had
special logic for processing subdomains. This can be removed now.