design: trim some of the extra padding off
In the GOV.UK component, there are 24px total padding above and below
the grid items. Trimming off some of that means the card list is
a bit more compact. Our design is a little more compact than the
original component as we want users to be able to see all the options
easily.