Small tweaks to the new homepage designs (#1301)
* design: remove redundant text from titles

* fix: tweak spacing of top item

If the padding is reduced for this item, then the height is shorter
and the arrow gets displayed a little closer to the seperator line.
This looks a bit off, so instead, use a negative margin to raise up
the first item.

* fix: add a little margin to the right of the cards

Otherwise the right edge of the cards looks a bit too close
to the "Help us grow" section.

* design: trim some of the extra padding off

In the GOV.UK component, there are 24px total padding above and below
the grid items. Trimming off some of that means the card list is
a bit more compact. Our design is a little more compact than the
original component as we want users to be able to see all the options
easily.

* build: create feature flags for the new homepage design

* design: remove bold text from links