fix: restore color of sign out link
The color of the navigation links is set to inherit, which should
mean white. However, if we blanket apply govuk list styles to all lists,
this interferes with that and makes it black, visually hiding it.

We should make this opt-in via a class on a parent container. As a quick
fix I've applied this to the main element, to catch all rendered markdown
on the details and glossary pages.