feat: highlight link when in that section of the glossary
The sticky menu on the left should make bold the section
that is currently in view.

When multiple are in view we should select the topmost one,
except when we are right at the bottom of the page, in which case we
should select the bottom one, so that the highlight moves all the way
from the top of the list to the bottom of the list.

The group determined to be currently in view changes based on a number
of interactions:
1. Clicking on a link
2. Scrolling
3. Resizing the page
4. Filtering and clearing the filter

To ensure that the highlight always matches whats visible, it is
recalculated on each of these events, but with throttling so we
don't overload the main thread.

I originally tried using the newer IntersectionObserver API for this[1],
but I found it didn't behave correctly when jumping around the page via
the anchor links.

[1]. https://developer.mozilla.org/en-US/docs/Web/API/Intersection_Observer_API