Glossary improvements (#1424)
* feat: add back the glossary link

* feat: add glossary term page

This page shows entities linked to the glossary term, with pagination.

Refactor the pagination component to a custom template tag, so that we
can customise the URL pattern used for the pagination links, rather than
hardcoding them as was done before.

* feat: expose glossary terms in search results and details pages

* feat: enable sane_lists in rendered markdown

Without this, the markdown renderer does not recognise nested lists.

* feat: glossary page improvements

Sidebar:
- remove links to individual terms,
  leaving just the term groups
- add spacing between the links
- add "On this page" heading
- add "Back to top" link

Main content:
- only show specific glossaries
- add a heading above the search box (but keep the existing visually
  hidden label, as it announces that the search results will change as
you type)
- remove the placeholder from search
- add an empty state for when there are no matches
- tweak spacing between glossary terms
- add back to top link
- link to glossary terms for Electronic monitoring terms (these are
  associated with entities in the catalogue)

* feat: highlight link when in that section of the glossary

The sticky menu on the left should make bold the section
that is currently in view.

When multiple are in view we should select the topmost one,
except when we are right at the bottom of the page, in which case we
should select the bottom one, so that the highlight moves all the way
from the top of the list to the bottom of the list.

The group determined to be currently in view changes based on a number
of interactions:
1. Clicking on a link
2. Scrolling
3. Resizing the page
4. Filtering and clearing the filter

To ensure that the highlight always matches whats visible, it is
recalculated on each of these events, but with throttling so we
don't overload the main thread.

I originally tried using the newer IntersectionObserver API for this[1],
but I found it didn't behave correctly when jumping around the page via
the anchor links.

[1]. https://developer.mozilla.org/en-US/docs/Web/API/Intersection_Observer_API

* fix: do not apply markdown fixes to glossary nav

* fixup: Show correct page title for glossary terms