Do not include the phase banner in landmarks
I originally added an aria-role of region to the phase banner because
axe complains if you have any elements on the page that don't belong
to a landmark.

> Ensure all content is contained within a landmark region, designated with HTML5 landmark elements and/or ARIA landmark regions.
>
> Screen reader users can navigate to a section based on its HTML element or ARIA Landmark. For example, you might use ARIA Landmarks to provide a simple replacement for a skip navigation link, though the replacement is only useful for users of screen readers. Sighted users or people using screen enlargers won't benefit from the addition, so it's not a good practice to substitute ARIA landmarks for skip navigation links altogether.

https://dequeuniversity.com/rules/axe/3.3/region?application=AxeChrome

However, from speaking to the GOV.UK design system team, it seems like
it's not recommended to create a distinct landmark just for the phase
banner, because it's not important enough. Early on GOV.UK made a
decision to reduce the number of landmarks to improve usability for
screenreader users (see
https://github.com/alphagov/govuk-frontend/issues/1604 for more
context).

The general advice is:
**If a banner can be reasonably added to any other landmark because it can
be interpreted to be part of it, or is at least somewhat related, it's
better to add it to that element rather than create a new landmark.**

Normally that would be the <header>, but in our case we display the
phase banner further down, below the nav, so I've included it in the
main element instead.

This has a minor drawback that the skiplink will not skip past the
phase banner, but I think it's better than the alternative, and
consistent with the phase banner variant of the x-gov masthead
component, which we are using on the homepage.